+++ changelog.md
@@ -1,3 +1,7 @@
+### 1.9-RC9 - 2020-10-01
+#### Fixed
+- Read Patient from endpoint '[GET] /patients/?tokenId={tokenid}]' cause an exception
+
 ### 1.9-RC8 - 2020-09-14
 #### Fixed
 - 500 Error caused by audit_trail feature if large requests where made with a ReadToken (Result > 255 Signs)